PR fixes (removed wall jump stub)
