--- Assets/GlitchEscape/Scripts/Player/Controls/PlayerJumpController_Alt.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62bdb20165c60414bb3bda67695eda65
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
