+++ Assets/GlitchEscape/Scripts/Player/Controls/PlayerPhysicsSettings.physicMaterial
@@ -7,8 +7,8 @@ PhysicMaterial:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: PlayerPhysicsSettings
-  dynamicFriction: 1
-  staticFriction: 0.5
+  dynamicFriction: 0.1
+  staticFriction: 0.1
   bounciness: 0
   frictionCombine: 0
   bounceCombine: 0
